project2/requirements.txt
Flask
Flask-SocketIO
sqlalchemy
werkzeug.security